Switch to using a 64bit value for shutdown time.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@1446957 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/threads/CompositeTaskRunner.cpp
@@ -73,7 +73,7 @@ bool CompositeTaskRunner::isStarted() const {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-void CompositeTaskRunner::shutdown(unsigned int timeout) {
+void CompositeTaskRunner::shutdown(long long timeout) {
 
     synchronized(&mutex) {
         shutDown = true;
